Handle seed control columns with a single value
I wanted to control the overall total households, so I created
a column in the seed table called total_hh and filled it with ones.
I created the accompanying target table. When run, the process
of creating dummy features failed - the function requires at least
two values in the column to work. This change handles that issue
by not converting the column to a factor but changing the column
name to what it needs to be.